Make vm.create able to take saner values for disksize creation.
Rather than having to work out size in KB you can now do:
"vm.create --size 50g ..." for a 50 Gigabyte disk. Values are in multiples
of 1024, rather than drive makers kilobytes.
Adjust defaults for the size parameter to be a string, and just rely on
the function to deal with it.