Add bootdelay boot options support.
Now possible to query and set the boot options for a VM.
This means you can change the delay between power on and it trying to boot
so you can get to the BIOS and boot off other media.
Also you can change the boot retry toggle and delay for when it fails to
find media. Which is useful in PXE environments.